requirements.txt
torch
torchvision
timm
transformers
peft
huggingface-hub
numpy
pandas
scikit-learn
h5py
openslide-python
openslide-bin
cucim
pillow
PyYAML
omegaconf
tqdm
wandb
scipy